requirements.txt
# Core dependencies for this project
SQLAlchemy==2.0.45
rich==14.2.0
qrcode==8.2
passlib==1.7.4
python-dotenv==1.0.1
